Download real-time, crowd-sourced weather data
This is a function ('Download_wunderground.py') and a call file
('weather_stations_list.py') that allow the user to download real-time,
crowd-sourced weather data from the WeatherUnderground weather station
network.

The user provides a unique site ID and a time period of data required
(start and end date) as strings. The obtained data is written to a text
file and saved in a specified directory for later, repeated access.

The function is a simple web-crawler application.